+++ src/ts-tooling.js
@@ -7,6 +7,11 @@ const {TimeSpan, DateTime, DateRange} = require('./types/datetime');
 const {Guid} = require('./types/guid');
 const {Byte, ByteStream} = require('./types/byte');
 const {Dictionary} = require('./types/dictionary');
+const {
+    FeatureCollection, Feature, Point, Line, Polygon,
+    MultiPoint, MultiLine, MultiPolygon,
+    ReferenceSystem, WGS84, WEB_MERCATOR, registerProjection
+} = require('./types/geometry');
 
 // Patterns
 const {Using, Create, CreateWithFactory, Throttle, Retry} = require('./pattern');