+++ src/types/geometry/index.d.ts
@@ -1,0 +1,12 @@
+export {
+    ReferenceSystem, ReferenceSystemProperties,
+    WEB_MERCATOR, WGS84, registerProjection,
+} from './reference.system';
+export {Point} from './point';
+export {Line} from './line';
+export {Polygon} from './polygon';
+export {MultiPoint} from './multi.point';
+export {MultiLine} from './multi.line';
+export {MultiPolygon} from './multi.polygon';
+export {Feature, IGeometry} from './feature';
+export {FeatureCollection} from './feature.collection';